[RA] Update Logic

+++ reliability_assessment/reliability_assessment.py
@@ -166,46 +166,46 @@ class ReliabilityAssessment:
         tweet_text = self.__remove_non_ascii(tweet['text'])
         return {'_id': tweet['_id'], 'output': self.nv_instance.detect(text=tweet_text, mode=mode)}
 
-    def neural_fake_news_detection(self):
+    def neural_fake_news_detection(self, gpt_2: bool, gltr: bool):
         # Always clean up fields before starting!
         if input('DO YOU WANT TO CLEAN RA RESULTS? (Y/N) ') == "Y":
             self.db_instance.remove_many('ra_raw', self.input_date, self.ticker)
 
-        self.nv_instance.init_gpt_model(model=DETECTOR_MAP['gpt-detector'])
-        # Split large tweets collection into smaller pieces -> GOOD FOR LAPTOP :)
-        SLICES = 10
-        for i in trange(0, len(self.tweets_collection), SLICES):
-            tweets_collection_small = self.tweets_collection[i:i + SLICES]
-            # Update RoBERTa-detector Results
-            with concurrent.futures.ThreadPoolExecutor() as executor:
-                gpt_2_futures = [executor.submit(self.detector_wrapper, tweet, 'gpt-2') for tweet in
-                                 tweets_collection_small]
+        if gpt_2:
+            self.nv_instance.init_gpt_model(model=DETECTOR_MAP['gpt-detector'])
+            # Split large tweets collection into smaller pieces -> GOOD FOR LAPTOP :)
+            SLICES = 10
+            for i in trange(0, len(self.tweets_collection), SLICES):
+                tweets_collection_small = self.tweets_collection[i:i + SLICES]
+                # Update RoBERTa-detector Results
+                with concurrent.futures.ThreadPoolExecutor() as executor:
+                    gpt_2_futures = [executor.submit(self.detector_wrapper, tweet, 'gpt-2') for tweet in
+                                     tweets_collection_small]
 
-            # Update MongoDB
-            for future in gpt_2_futures:
-                self.db_instance.update_one(future.result()['_id'], 'ra_raw.RoBERTa-detector',
-                                            future.result()['output'],
-                                            self.input_date, self.ticker)
+                # Update MongoDB
+                for future in gpt_2_futures:
+                    self.db_instance.update_one(future.result()['_id'], 'ra_raw.RoBERTa-detector',
+                                                future.result()['output'],
+                                                self.input_date, self.ticker)
+            # Kill GPT-2 Process
+            result = [p.kill() for p in SUB_PROCESSES]
 
-        # Kill GPT-2 Process
-        result = [p.kill() for p in SUB_PROCESSES]
-
-        self.nv_instance.init_gltr_models(models=DETECTOR_MAP['gltr-detector'])
-        SLICES = 3
-        for i in trange(0, len(self.tweets_collection), SLICES):
-            tweets_collection_small = self.tweets_collection[i:i + SLICES]
-            # Update GLTR Results
-            with concurrent.futures.ThreadPoolExecutor() as executor:
-                gltr_futures = [executor.submit(self.detector_wrapper, tweet, 'gltr') for tweet in
-                                tweets_collection_small]
-            for future in gltr_futures:
-                self.db_instance.update_one(future.result()['_id'],
-                                            f"ra_raw.{DETECTOR_MAP['gltr-detector'][0]}-detector",
-                                            future.result()['output'][0], self.input_date, self.ticker)
-                self.db_instance.update_one(future.result()['_id'],
-                                            f"ra_raw.{DETECTOR_MAP['gltr-detector'][1]}-detector",
-                                            future.result()['output'][1], self.input_date, self.ticker)
-
-        result = [p.kill() for p in SUB_PROCESSES]
+        if gltr:
+            self.nv_instance.init_gltr_models(models=DETECTOR_MAP['gltr-detector'])
+            SLICES = 3
+            for i in trange(0, len(self.tweets_collection), SLICES):
+                tweets_collection_small = self.tweets_collection[i:i + SLICES]
+                # Update GLTR Results
+                with concurrent.futures.ThreadPoolExecutor() as executor:
+                    gltr_futures = [executor.submit(self.detector_wrapper, tweet, 'gltr') for tweet in
+                                    tweets_collection_small]
+                for future in gltr_futures:
+                    self.db_instance.update_one(future.result()['_id'],
+                                                f"ra_raw.{DETECTOR_MAP['gltr-detector'][0]}-detector",
+                                                future.result()['output'][0], self.input_date, self.ticker)
+                    self.db_instance.update_one(future.result()['_id'],
+                                                f"ra_raw.{DETECTOR_MAP['gltr-detector'][1]}-detector",
+                                                future.result()['output'][1], self.input_date, self.ticker)
+            result = [p.kill() for p in SUB_PROCESSES]
 
         self.default_logger.info("Neural Fake News Detector Output Update Success! ")
